refactor(dx): Simplify adapter setup and improve DX
This commit enhances the developer experience by simplifying the setup for both React and Vue adapters.

- **React**: RouterProvider now renders <RouterOutlet /> by default when no children are provided, removing the need for explicit nesting in main.tsx.
- **Vue**: createApp can now be used directly with RouterOutlet, streamlining the application entry point and eliminating the need for a separate App.vue wrapper.
- **Vue**: The useRouterContext composable is now co-located with other composables in src/adapters/vue/composables.ts for better code organization.
- **Docs**: Documentation and examples have been updated to reflect these DX improvements.
