--- playground/vue3/src/App.vue
@@ -1,7 +1,0 @@
-<template>
-  <RouterOutlet />
-</template>
-
-<script setup lang="ts">
-import { RouterOutlet } from '@ciderjs/city-gas/vue';
-</script>
