+++ playground/vue3/src/main.ts
@@ -1,12 +1,11 @@
 import { createRouter } from '@ciderjs/city-gas';
-import { createRouterPlugin } from '@ciderjs/city-gas/vue';
+import { createRouterPlugin, RouterOutlet } from '@ciderjs/city-gas/vue';
 import { createApp } from 'vue';
-import App from './App.vue';
 import { pages, specialPages } from './generated/routes';
 
 function main() {
   const router = createRouter(pages, { specialPages });
-  createApp(App).use(createRouterPlugin(router)).mount('#root');
+  createApp(RouterOutlet).use(createRouterPlugin(router)).mount('#root');
 }
 
 main();
