+++ src/adapters/react/provider.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { RouterOutlet } from '@/adapters/react/RouterOutlet';
 import type { Router } from '@/core/router';
 
 export const RouterContext = React.createContext<{
@@ -10,13 +11,13 @@ export function RouterProvider<R extends string, P extends Record<R, any>>({
   children,
 }: {
   router: Router<R, P>;
-  children: React.ReactNode;
+  children?: React.ReactNode;
 }) {
   const contextValue = React.useMemo(() => ({ router }), [router]);
 
   return (
     <RouterContext.Provider value={contextValue}>
-      {children}
+      {children ?? <RouterOutlet />}
     </RouterContext.Provider>
   );
 }
