+++ src/adapters/vue/index.ts
@@ -3,6 +3,7 @@ export {
   useParams,
   useRoute,
   useRouter,
+  useRouterContext,
 } from '@/adapters/vue/composables';
 export { createRouterPlugin } from '@/adapters/vue/plugin';
 export { RouterOutlet } from '@/adapters/vue/RouterOutlet';
